requirements.txt
asgiref==3.4.1
Django==4.0.1
django-leaflet==0.28.2
djangorestframework==3.13.1
geographiclib==1.52
geopy==2.2.0
Pillow==9.0.0
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
